That's all, folks.
