+++ app/entry.server.tsx
@@ -1,17 +1,14 @@
 import { PassThrough } from "node:stream"
 import { createReadableStreamFromReadable } from "@react-router/node"
-import type { Context } from "hono"
 import { createInstance } from "i18next"
 import { isbot } from "isbot"
 import { renderToPipeableStream } from "react-dom/server"
 import { I18nextProvider, initReactI18next } from "react-i18next"
 import { type AppLoadContext, type EntryContext, ServerRouter } from "react-router"
-import { createHonoServer } from "react-router-hono-server/node"
-import { i18next } from "remix-hono/i18next"
-import { getClientEnv, initEnv } from "./env.server"
 import i18n from "./localization/i18n" // your i18n configuration file
 import i18nextOpts from "./localization/i18n.server"
 import { resources } from "./localization/resource"
+
 const ABORT_DELAY = 5000
 
 export default async function handleRequest(
@@ -73,41 +70,3 @@ export default async function handleRequest(
 		setTimeout(abort, ABORT_DELAY)
 	})
 }
-
-// Code below used to initialize our own Hono server!
-// Setup the .env vars
-const env = initEnv()
-
-const getLoadContext = async (c: Context) => {
-	// get the locale from the context
-	const locale = i18next.getLocale(c)
-	// get t function for the default namespace
-	const t = await i18next.getFixedT(c)
-
-	const clientEnv = getClientEnv()
-	return {
-		lang: locale,
-		t,
-		env,
-		clientEnv,
-		// We do not add this to AppLoadContext type because it's not needed in the loaders, but it's used above to handle requests
-		body: c.body,
-	}
-}
-
-interface LoadContext extends Awaited<ReturnType<typeof getLoadContext>> {}
-
-/**
- * Declare our loaders and actions context type
- */
-declare module "react-router" {
-	interface AppLoadContext extends Omit<LoadContext, "body"> {}
-}
-
-export const server = await createHonoServer({
-	configure(server) {
-		server.use("*", i18next(i18nextOpts))
-	},
-	defaultLogger: false,
-	getLoadContext,
-})
