+++ package.json
@@ -11,9 +11,9 @@
 		"clean": "git clean -fdX --exclude=\"!.env\"",
 		"script": "tsx scripts/setup.ts",
 		"cleanup": "pnpm run script scripts/cleanup.ts",
-		"build": "NODE_ENV=production react-router build",
+		"build": "react-router build",
 		"dev": "react-router dev",
-		"start": "NODE_ENV=production node ./build/server/index.js",
+		"start": "NODE_ENV=production node --env-file .env ./build/server/index.js",
 		"test": "vitest run",
 		"test:ui": "vitest --ui --api 9527",
 		"test:cov": "vitest run --coverage",
@@ -30,7 +30,7 @@
 		"@forge42/seo-tools": "1.3.0",
 		"@react-router/node": "7.0.1",
 		"clsx": "2.1.1",
-		"hono": "4.6.11",
+		"hono": "4.6.12",
 		"i18next": "23.15.2",
 		"i18next-browser-languagedetector": "8.0.0",
 		"i18next-fetch-backend": "6.0.0",
@@ -40,7 +40,7 @@
 		"react-dom": "18.3.1",
 		"react-i18next": "15.1.1",
 		"react-router": "7.0.1",
-		"react-router-hono-server": "https://pkg.pr.new/rphlmr/react-router-hono-server@18.tgz",
+		"react-router-hono-server": "https://pkg.pr.new/rphlmr/react-router-hono-server@18",
 		"remix-hono": "0.0.16",
 		"remix-i18next": "7.0.0",
 		"tailwind-merge": "2.5.4",
